GUI Update 26 external DB paremeters
Work continues on the external DB parameters view in the edit portion of the menu. Testing works against hard coded db parameters in the code (too difficult to keep typing them in), next up is getting export to work. I'm using my own sparterASP account to test against, the db values are in the editviewmodel code.